requirements.txt
Django==1.8
mysqlclient==1.3.12
python-dateutil==2.6.1
pytz==2017.2
